# dist/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Coppia.co API Documentation",
        "description": "API Documentation for Coppia.co API calls",
        "termsOfService": "http://swagger.io/terms/",
        "contact" : {
            "name" : "Coppia API Support",
            "url" : "www.coppia.co/support",
            "email" : "eric@coppia.co"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
...
    "schemes": [
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/api/v1/interviews/": {
            "get": {
                "tags": ["Interviews"],
                "description": "Return list of Interviews",
                "operationId": "find",
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Interview object returned in list",
                        "schema": {
                            "$ref": "#/definitions/interview"
                        }
                    }
                }
            },
            "post":{
                "tags": ["Interviews"],
                "description": "Creates a new Inteview",
                "operationId": "create",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "title",
                        "in": "body",
                        "description": "Title for the Interview",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "notes",
                        "in": "body",
                        "description": "Notes for the Interview",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "create_user",
                        "in": "body",
                        "description": "User Id for the logged in user",
                        "required": true,
                        "type": "integer",
                        "format" : "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "When creating a new interview, the id for the newly created Interview will be returned",
                        "schema": {
                            "$ref": "#/definitions/create_interview"
                        }
                    }
                }
            }
        },
        "/api/v1/interviews/{interview_id}": {
            "get":{
                "tags": ["Interviews"],
                "description": "Returns a single Interview object",
                "operationId": "findById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "interview_id",
                        "in": "path",
                        "description": "id for the Interview user wants to get",
                        "required": true,
                        "type": "integer",
                        "format" : "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns single Interview object",
                        "schema": {
                            "$ref": "#/definitions/interview"
                        }
                    }
                }
            },
            "put":{
                "tags": ["Interviews"],
                "description": "Returns a success message with count of interviews updates",
                "operationId": "updateById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "interview_id",
                        "in": "path",
                        "description": "id for the Interview to update",
                        "required": true,
                        "type": "integer",
                        "format" : "int32"
                    },
                    {
                        "name": "title",
                        "in": "body",
                        "description": "title for the Interview",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "notes",
                        "in": "body",
                        "description": "notes for the Interview",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "update_user",
                        "in": "body",
                        "description": "id for the logged in user",
                        "required": true,
                        "type": "integer",
                        "format" : "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response when calling to update an Interview",
                        "schema": {
                            "$ref": "#/definitions/update_interview"
                        }
                    }
                }
            },
            "delete":{
                "tags":["Interviews"],
                "description":"Deletes an interview from the database",
                "operationId":"delete",
                "produces": [
                    "application/json"
                ],
                "parameters":[
                    {
                        "name": "interview_id",
                        "in": "path",
                        "description": "interview_id for the Interview to be deleted from the database. This will remove the association in the inteview_customer table as well",
                        "required": true,
                        "type": "integer",
                        "format" : "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response for the deleted Interview",
                        "schema":{
                            "$ref":"#/definitions/delete_interview"
                        }
                    }
                }
            }
        },
        "/api/v1/interviews/interview_customer/{interview_id}": {
            "get":{
                "tags": ["Interviews"],
                "description": "Returns a single Customer object",
                "operationId": "findCustomerByInterviewId",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "interview_id",
                        "in": "path",
                        "description": "id for the Interview associated to the customer",
                        "required": true,
                        "type": "integer",
                        "format" : "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns single Customer object",
                        "schema": {
                            "$ref": "#/definitions/customer"
                        }
                    }
                }
            }
        },
        "/api/v1/interviews/interview_customer/": {
            "post":{
                "tags": ["Interviews"],
                "description": "Creates a new interview_customer association",
                "operationId": "create",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "interview_id",
                        "in": "body",
                        "description": "id for the interview to associate",
                        "required": true,
                        "type": "integer",
                        "format" : "int32"
                    },
                    {
                        "name": "customer_id",
                        "in": "body",
                        "description": "id for the customer to associate",
                        "required": true,
                        "type": "integer",
                        "format" : "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "When creating a new interview_customer association, the id for the newly created intervew_customer will be returned",
                        "schema": {
                            "$ref": "#/definitions/create_interview_customer"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "interview": {
            "required": [
                "id",
                "title",
                "notes",
                "create_user",
                "create_datetime",
                "update_user",
                "update_datetime"
            ],
            "properties": {
                "interview_id": {
                    "type": "integer",
                    "format" : "int32",
                    "example": "6"
                },
                "title": {
                    "type": "string",
                    "example": "Reporting Discovery w/Charlotte"
                },
                "notes": {
                    "type": "string",
                    "example": "Industrial distributor for padlocks. Padlocks are very custom items. Shackle lengths, master keys, a lot of customization that goes into ordering padlocks. Can you explain this keying to me? Using chat would help reduce the phone traffic. Those conversations are relatively quick. They have 5 people with the company. They all take turns answering calls and 3 people are on Pure Chat every day."
                },
                "created_by": {
                    "type": "string",
                    "example": "Eric Hillenbrand"
                },
                "created_datetime": {
                    "type": "string",
                    "format" : "date-time",
                    "example": "2016-10-17T23:10:23.000Z"
                },
                "created_image_link": {
                    "type": "string",
                    "example": "https://d1ts43dypk8bqh.cloudfront.net/v1/avatars/88a1184e-964b-495b-86fc-92c2867c5427"
                },
                "updated_by": {
                    "type": "string",
                    "example": "Eric Hillenbrand"
                },
                "updated_datetime": {
                    "type": "string",
                    "format" : "date-time",
                    "example": "2016-10-17T23:10:23.000Z"
                },
                 "updated_image_link": {
                    "type": "string",
                    "example": "https://d1ts43dypk8bqh.cloudfront.net/v1/avatars/88a1184e-964b-495b-86fc-92c2867c5427"
                }
            }
        },
        "create_interview": {
            "properties": {
                "interview_id": {
                    "type": "integer",
                    "format" : "int32",
                    "example" : "3"
                }
            }
        },
         "update_interview": {
            "properties": {
                "interview_id": {
                    "type": "integer",
                    "format" : "int32",
                    "example" : "8"
                }
            }
        },
         "delete_interview": {
            "properties": {
                "interview_id": {
                    "type": "integer",
                    "format" : "int32",
                    "example" : "6"
                }
            }
        },
        "create_interview_customer": {
            "properties": {
                "interview_customer_id": {
                    "type": "integer",
                    "format" : "int32",
                    "example" : "44"
                }
            }
        },
        "customer": {
            "required": [
                "id",
                "first_name",
                "last_name",
                "email"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "format" : "int32",
                    "example": "6"
                },
                "first_name": {
                    "type": "string",
                    "example": "Eric"
                },
                "last_name": {
                    "type": "string",
                    "example": "Hillenbrand"
                },
                "email": {
                    "type": "string",
                    "example": "eric@coppia.co"
                },
                "image_link": {
                    "type": "string",
                    "example": "https://d1ts43dypk8bqh.cloudfront.net/v1/avatars/88a1184e-964b-495b-86fc-92c2867c5427"
                },
                "title": {
                    "type": "string",
                    "example": "Product Marketing Manager"
                },
                "company_name": {
                    "type": "string",
                    "example": "Web PT"
                },
                "created_by": {
                    "type": "string",
                    "example": "Eric Hillenbrand"
                },
                "created_date": {
                    "type": "string",
                    "format" : "date-time",
                    "example": "2016-10-17T23:10:23.000Z"
                },
                "created_image_link": {
                    "type": "string",
                    "example": "https://d1ts43dypk8bqh.cloudfront.net/v1/avatars/88a1184e-964b-495b-86fc-92c2867c5427"
                },
                "updated_by": {
                    "type": "string",
                    "example": "Eric Hillenbrand"
                },
                "updated_datetime": {
                    "type": "string",
                    "format" : "date-time",
                    "example": "2016-10-17T23:10:23.000Z"
                },
                "updated_image_link": {
                    "type": "string",
                    "example": "https://d1ts43dypk8bqh.cloudfront.net/v1/avatars/88a1184e-964b-495b-86fc-92c2867c5427"
                }
            }
        },
...
    }
}